# pizzaSiteAppModules/api/src/main/resources/openapi/new/0420.yaml
openapi: 3.0.0
info:
  title: Pizza project API
  version: 1.1.0
  description: |
    Pizza API allows you to remotely manage the pizzeria.
    ### Description of the project and requirements:
    * Authorization is required to work with the API. Authorization is done by the bearer-token 😎
servers:
  - url: http://localhost:8090/
paths:
...
  /api/users/all:
    get:
      summary: Get all users
      tags:
        - Users
      operationId: get all users
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Person'
        '401':
          description: Access denied
  /api/users/{id}:
    get:
      summary: Get user
      description: Get one by id
      tags:
        - Users
      operationId: get user
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200' :
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'

        '401':
          description: Access denied
        '404':
          description: Request failed - No items
  /api/users/edit:
    patch:
      summary: Change users info
      tags:
        - Users
      operationId: update user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                email:
                  type: string
              required:
                - username
                - password
                - email
      responses:
        200 :
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    jwt-token:
                      type: string
                      example: 'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJVc2VyIGRldGFpbHMiLGHpc3MiOiJQaXp6YSBhcHAiLCJleHAiOjE2ODE2MzU5NzgsImlhdCI6MTY4MTYzMjM3OCwidXNlcm5hbWUiOiJ0ZXN0In0.XvbElw5HT59q8aEPIEOm7wgbuIMou3_WfcogK4gJ__I'
                    updated data user:
                      type: string
                      example: 'newUserName : newUserEmail'
        404:
          description: Invalid data
        401:
          description: Access denied
...
  /api/users/addToFav/{id}/{pizzaId}:
    post:
      summary: Add pizza to users Favorites list
      description: Add pizza
      tags:
        - Users
      operationId: add to Fav
      parameters:
        - $ref: "#/components/parameters/id"
        - $ref: "#/components/parameters/pizzaId"
      responses:
        '200' :
          description: Successful operation
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: "Pizza id 1 was added to favorites user id 1"
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: "description"
        '404':
          description: Request failed - No items
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: "description"
  /api/users/removeFromFav/{id}/{pizzaId}:
    post:
      summary: Remove pizza from users Favorites list
      description: Remove pizza
      tags:
        - Users
      operationId: remove from Fav list
      parameters:
        - $ref: "#/components/parameters/id"
        - $ref: "#/components/parameters/pizzaId"
      responses:
        '200' :
          description: Successful operation
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: "Pizza id 1 was removed from favorites user id 1"
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: "description"
        '404':
          description: Request failed - No items
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: "description"
...
  /api/admin/pizza/change/{pizzaId}:
    post:
      summary: Change pizza (Admin only)
      description: Change pizza
      operationId: change pizza
      tags:
        - Admin
      parameters:
        - $ref: '#/components/parameters/pizzaId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BodyPizza'
      responses:
        '200' :
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pizza'
        '403':
          description: Access denied
          content:
            appllication/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Only admins are allowed access"
        '404':
          description: Incorrect data
          content:
            appllication/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "The data entered is not correct"
  /api/admin/ingredient/change/{id}:
    post:
      summary: Change ingredient (Admin only)
      description: Change ingredient
      operationId: change ingredient
      tags:
        - Admin
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BodyIngredient'
      responses:
        '200' :
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ingredient'
        '403':
          description: Access denied
          content:
            appllication/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Only admins are allowed access"
        '404':
          description: Incorrect data
          content:
            appllication/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "The data entered is not correct"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      in: path
      name: id
      required: true
      description: record id
      schema:
        type: integer
    pizzaId:
      in: path
      name: pizzaId
      required: true
      description: pizza id
      schema:
        type: integer
...
  schemas:
    Person:
      type: object
      required:
        - username
        - password
        - email
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: "John Smith"
        password:
          type: string
        email:
          type: string
          example: "test@email.com"
...
    Base:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'Thin crust'
        size:
          type: string
          example: 'Small'
        price:
          type: number
          example: '3.50'
...
    TypeOfInredient:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'Meat'
    Ingredient:
      type: object
      properties:
        id:
          type: integer
          example: 32
        name:
          type: string
          example: 'Bacon'
        price:
          type: number
          example: '4.50'
        image:
          type: string
          example: 'i_bacon.jpg'
        type:
          $ref: '#/components/schemas/TypeOfInredient'
    BodyIngredient:
      type: object
      properties:
        name:
          type: string
          example: 'Royal Cheese'
        price:
          type: number
          example: '3.50'
        type_id:
          type: integer
          example: 3
        image:
          type: string
          example: 'royal_chesse.jpg'
      required:
        - name
        - price
        - type_id
    Pizza:
      type: object
      properties:
        id:
          type: integer
          example: 2
        name:
          type: string
          example: 'Margarita S'
        price:
          type: number
          example: '15.50'
        base:
          $ref: '#/components/schemas/Base'
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/Ingredient'
        image:
          type: string
          example: 'p_margarita.png'
    BodyPizza:
      type: object
      properties:
        name:
          type: string
          example: 'Margarita S'
        price:
          type: number
          example: 25.50
        base_id:
          type: integer
          example: 2
        image:
          type: string
          example: 'p_margarita.png'
      required:
        - name
        - price
        - base_id
...
security:
  - bearerAuth: []